Fixed java 8 issue with test

+++ src/tests/net/sf/persist/tests/performance/PerformanceH2.java
@@ -26,7 +26,7 @@ import com.jamonapi.Monitor;
 import com.jamonapi.MonitorFactory;
 
 public class PerformanceH2 {
-	
+
 	public static AllTypes buildAllTypes() {
 		AllTypes a = new AllTypes();
 		byte[] binaryCol = new byte[255]; for (int i=0; i<255; i++) binaryCol[i]=(byte)(i/2);
@@ -53,39 +53,39 @@ public class PerformanceH2 {
 	}
 
 	public static void main(String[] args) throws SQLException, IOException {
-		
+
 		String reportPath = args[0];
-		
+
 		String properties = "net/sf/persist/tests/performance/h2.properties";
-		
+
 		// create a connection
 		Connection connection = ConnectionHelper.getConnection(properties);
 		connection.setAutoCommit(true);
 		connection.createStatement().execute("create table if not exists all_types ( int_col int, boolean_col boolean, tinyint_col tinyint, smallint_col smallint, bigint_col bigint, decimal_col decimal, double_col double, real_col real, time_col time, date_col date, timestamp_col timestamp, binary_col binary, blob_col blob, other_col other, uuid_col uuid, varchar_col varchar, varchar_ignorecase_col varchar_ignorecase, char_col char, clob_col clob, id int auto_increment )");
 		connection.createStatement().execute("delete from all_types");
-		
+
 		// persist
-		Persist persist = new Persist(connection);		
-		
+		Persist persist = new Persist(connection);
+
 		int inserts = 5000;
 		int reads = 10000;
 		int readLists = 5;
-		
+
 		AllTypes a = buildAllTypes();
-		
+
 		// insert
-		
+
 		System.out.print("insert");
 
 		String sql = "insert into ALL_TYPES(INT_COL,BOOLEAN_COL,TINYINT_COL,SMALLINT_COL,BIGINT_COL,DECIMAL_COL," +
 		"DOUBLE_COL,REAL_COL,TIME_COL,DATE_COL,TIMESTAMP_COL,BINARY_COL,BLOB_COL,OTHER_COL,UUID_COL," +
 		"VARCHAR_COL,VARCHAR_IGNORECASE_COL,CHAR_COL,CLOB_COL)values(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
-		
+
 		for (int i=1; i<=inserts; i++) {
-			
+
 			if (i%10==0) System.out.print(" " + i);
 			if (i%250==0) System.out.println();
-			
+
 			// jdbc
 			Monitor mon = MonitorFactory.start("[insert] jdbc");
 			PreparedStatement stmt = connection.prepareStatement(sql);
@@ -111,29 +111,29 @@ public class PerformanceH2 {
 			stmt.executeUpdate();
 			stmt.close();
 			mon.stop();
-			
+
 			// persist
 			mon = MonitorFactory.start("[insert] persist");
 			persist.insert(a);
 			mon.stop();
-			
+
 		}
 		System.out.println();
-		
-		
+
+
 		// find a single id to be used by select by primary key
-		long id = persist.read(int.class, "select min(id) from all_types");
-		
+		long id = persist.read(int.class, "select min(id) from all_types").longValue();
+
 		// read
-		
+
 		System.out.print("read");
-		
+
 		a = new AllTypes();
-		
+
 		for (int i=1; i<=reads; i++) {
 			if (i%50==0) System.out.print(" " + i);
 			if (i%1000==0) System.out.println();
-			
+
 			// jdbc
 			Monitor mon = MonitorFactory.start("[read single] jdbc");
 			PreparedStatement stmt = connection.prepareStatement("select * from all_types where id=?");
@@ -161,22 +161,22 @@ public class PerformanceH2 {
 			a.setClobCol(rs.getString(19));
 			stmt.close();
 			mon.stop();
-			
+
 			mon = MonitorFactory.start("[read single] persist");
 			a = persist.readByPrimaryKey(AllTypes.class, id);
 			mon.stop();
-			
+
 			mon = MonitorFactory.start("[read single] persist map");
 			/*Map<String, Object> m =*/ persist.readMap("select * from all_types where id=?", id);
 			mon.stop();
 		}
 		System.out.println();
-		
+
 		System.out.print("read list");
 		for (int i=1; i<=readLists; i++) {
 			if (i%1==0) System.out.print(" " + i);
 			if (i%20==0) System.out.println();
-			
+
 			List al = new ArrayList();
 			Monitor mon = MonitorFactory.start("[read list] jdbc");
 			PreparedStatement stmt = connection.prepareStatement("select * from all_types");
@@ -209,16 +209,16 @@ public class PerformanceH2 {
 			mon = MonitorFactory.start("[read list] persist");
 			/*List<AllTypes> list =*/ persist.readList(AllTypes.class);
 			mon.stop();
-			
+
 			mon = MonitorFactory.start("[read list] persist map");
 			/*List<Map<String,Object>> map =*/ persist.readMapList("select * from all_types");
 			mon.stop();
 		}
 		System.out.println();
-		
+
 		System.out.println();
 		System.out.println("finished");
-		
+
 		// read sort table javascript
 		InputStream is = Thread.currentThread().getContextClassLoader().getResourceAsStream("net/sf/persist/tests/performance/sorttable.js");
 		BufferedReader r = new BufferedReader(new InputStreamReader(is));
@@ -227,7 +227,7 @@ public class PerformanceH2 {
 		while ((line = r.readLine())!=null) {
 			sorttable.append(line+"\n");
 		}
-		
+
 		// create report html
 		System.out.println("writing report to " + reportPath);
 		String report = MonitorFactory.getReport();
@@ -238,7 +238,7 @@ public class PerformanceH2 {
 		w.write(report.replace("<table", "<table id=\"report-table\""));
 		w.write("</body></html>");
 		w.close();
-		
+
 	}
 
 }
